(MAINT) fix connection preferences
Previously, beaker-vmpooler defined its
connection preferences as string options, but
beaker provides defaults as symbols. This ends
with beaker saying that it doesn't support methods
that it does due to the mismatch. This will get
rid of those messages, as the provided methods
will be symbols, matching what beaker expects